+++ topi/python/topi/__init__.py
@@ -9,7 +9,29 @@ specific workload.
 """
 from __future__ import absolute_import as _abs
 
-from ._ffi.base import TVMError, __version__
+import sys
+import os
+import ctypes
+from tvm._ffi import libinfo
+
+def get_lib_names():
+    if sys.platform.startswith('win32'):
+        return ['libtvm_topi.dll', 'tvm_topi.dll']
+    if sys.platform.startswith('darwin'):
+        return ['libtvm_topi.dylib', 'tvm_topi.dylib']
+    return ['libtvm_topi.so', 'tvm_topi.so']
+
+curr_path = os.path.dirname(os.path.abspath(os.path.expanduser(__file__)))
+lib_search = curr_path
+
+def _load_lib():
+    """Load libary by searching possible path."""
+    lib_path = libinfo.find_lib_path(get_lib_names(), lib_search)
+    lib = ctypes.CDLL(lib_path[0], ctypes.RTLD_GLOBAL)
+    return lib, os.path.basename(lib_path[0])
+
+__version__ = libinfo.__version__
+_LIB, _LIB_NAME = _load_lib()
 
 from .math import *
 from .reduction import *
