+++ topi/python/topi/_ffi/__init__.py
@@ -1,5 +1,2 @@
 # pylint: disable=wildcard-import
 """FFI implementation"""
-from __future__ import absolute_import as _abs
-
-from .cpp_init import *
